pyproject.toml
[tool.poetry]
name = "campaigneon"
version = "0.1.0"
description = "An application for managing and generating content for DnD campaigns."
authors = ["anderskloborg <anderskloborg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
openai = "^0.28.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-dotenv = "^1.0.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.209.5"}
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
asyncpg = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
